1.院校收藏统计开发

+++ gk-user-web/src/main/java/cn/thinkjoy/gk/controller/UserCollectController.java
@@ -8,6 +8,7 @@ import cn.thinkjoy.gk.pojo.UserAccountPojo;
 import cn.thinkjoy.gk.pojo.UserCollectPojo;
 import cn.thinkjoy.gk.pojo.VideoCoursePojo;
 import cn.thinkjoy.gk.protocol.ERRORCODE;
+import cn.thinkjoy.gk.service.ICountExService;
 import cn.thinkjoy.gk.service.IUserCollectExService;
 import cn.thinkjoy.gk.service.IUserCollectService;
 import com.google.common.collect.Maps;
@@ -37,6 +38,9 @@ public class UserCollectController extends BaseController{
     @Autowired
     private IUserCollectExService userCollectExService;
 
+    @Autowired
+    private ICountExService countExService;
+
     /**
      * 保存当前用户的收藏项目（学校或课程）
      * @param projectId
@@ -44,7 +48,7 @@ public class UserCollectController extends BaseController{
      */
     @RequestMapping(value = "/saveUserCollect",method = RequestMethod.GET)
     @ResponseBody
-    public boolean saveUserCollect(@RequestParam(value = "projectId",required = true)long projectId){
+    public boolean saveUserCollect(@RequestParam(value = "projectId",required = true)long projectId,@RequestParam(value = "type",required = true)String type){
         UserAccountPojo userAccountPojo=getUserAccountPojo();
         if(null==userAccountPojo ||  null==userAccountPojo.getId()){
             throw new BizException(ERRORCODE.USER_NO_EXIST.getCode(), ERRORCODE.USER_NO_EXIST.getMessage());
@@ -53,17 +57,19 @@ public class UserCollectController extends BaseController{
         Map<String,Object> param=Maps.newHashMap();
         param.put("userId",userId);
         param.put("projectId",projectId);
-        if(userCollectExService.isUniversityCollect(param)>0){
+        param.put("type",type);
+        if(userCollectExService.isCollect(param)>0){
             throw new BizException(ERRORCODE.FAIL.getCode(),ERRORCODE.FAIL.getMessage());
         }
         UserCollect userCollect=new UserCollect();
         userCollect.setUserId(userId);
         userCollect.setProjectId(projectId);
-        userCollect.setType("1");
+        userCollect.setType(type);
         userCollect.setCreateDate(System.currentTimeMillis());
         boolean flag = false;
         try {
             userCollectService.add(userCollect);
+            countExService.updateCount(projectId, type,"add");
             flag = true;
         }catch(Exception e){
             e.printStackTrace();
@@ -82,18 +88,18 @@ public class UserCollectController extends BaseController{
     @ResponseBody
     public boolean deleteUserCollect(@RequestParam(value = "projectId",required = true)long projectId,@RequestParam(value = "type",required = true)String type){
         boolean flag = false;
-//        UserAccountPojo userAccountPojo=getUserAccountPojo();
-//        if(null==userAccountPojo ||  null==userAccountPojo.getId()){
-//            throw new BizException(ERRORCODE.USER_NO_EXIST.getCode(), ERRORCODE.USER_NO_EXIST.getMessage());
-//        }
-//        long userId=userAccountPojo.getId();
-        long userId=1l;
+        UserAccountPojo userAccountPojo=getUserAccountPojo();
+        if(null==userAccountPojo ||  null==userAccountPojo.getId()){
+            throw new BizException(ERRORCODE.USER_NO_EXIST.getCode(), ERRORCODE.USER_NO_EXIST.getMessage());
+        }
+        long userId=userAccountPojo.getId();
         Map<String,Object> param=Maps.newHashMap();
-        param.put("userId",userId);
+        param.put("userId", userId);
         param.put("projectId", projectId);
         param.put("type", type);
         try {
             userCollectService.deleteByCondition(param);
+            countExService.updateCount(projectId, type,"reduce");
             flag = true;
         }catch (Exception e){
             e.printStackTrace();
@@ -103,14 +109,14 @@ public class UserCollectController extends BaseController{
     }
 
     /**
-     * 查询当前用户的收藏院校
+     * 查询当前用户收藏的特定项目列表
      * @param offset
      * @param rows
      * @return
      */
-    @RequestMapping(value = "/getUserCollectUniversityPojoList",method = RequestMethod.GET)
+    @RequestMapping(value = "/getUserCollectList",method = RequestMethod.GET)
     @ResponseBody
-    public Map<String,Object> getUserCollectUniversityPojoList(@RequestParam(value ="offset",required = false)Integer offset, @RequestParam(value = "rows",required = false)Integer rows){
+    public Map<String,Object> getUserCollectList(@RequestParam(value ="type",required = true)String type,@RequestParam(value ="offset",required = false)Integer offset, @RequestParam(value = "rows",required = false)Integer rows){
         UserAccountPojo userAccountPojo=getUserAccountPojo();
         if(null==userAccountPojo ||  null==userAccountPojo.getId()){
             throw new BizException(ERRORCODE.USER_NO_EXIST.getCode(), ERRORCODE.USER_NO_EXIST.getMessage());
@@ -120,11 +126,22 @@ public class UserCollectController extends BaseController{
         param.put("userId",userId);
         param.put("offset",offset);
         param.put("rows",rows);
-        int sum=userCollectExService.getUserCollectUniversityPojoCount(param);
-        List<UserCollectPojo> userCollectPojoList=userCollectExService.getUserCollectUniversityPojoList(param);
-        Map<String,Object> returnParam=Maps.newHashMap();
-        returnParam.put("sum",sum);
-        returnParam.put("userCollectPojoList",userCollectPojoList);
+        param.put("type",type);
+        Map<String, Object> returnParam = Maps.newHashMap();
+        //收藏学校
+        if(type.equals("1")) {
+            int sum = userCollectExService.getUserCollectCount(param);
+            List<UserCollectPojo> userCollectPojoList = userCollectExService.getUserCollectUniversityPojoList(param);
+            returnParam.put("sum", sum);
+            returnParam.put("userCollectPojoList", userCollectPojoList);
+        }
+        //收藏课程
+        else if(type.equals("2")){
+            List<VideoCoursePojo> videoCoursePojoList=userCollectExService.getUserCollectVideoCoursePojoList(param);
+            int sum = userCollectExService.getUserCollectCount(param);
+            returnParam.put("sum", sum);
+            returnParam.put("videoCoursePojoList",videoCoursePojoList);
+        }
         return returnParam;
     }
 
@@ -135,7 +152,7 @@ public class UserCollectController extends BaseController{
      */
     @RequestMapping(value = "/isUniversityCollect",method = RequestMethod.GET)
     @ResponseBody
-    public int isUniversityCollect(@RequestParam(value = "projectId")String projectId){
+    public int isUniversityCollect(@RequestParam(value = "projectId",required = true)String projectId,@RequestParam(value = "type",required = true)String type){
         UserAccountPojo userAccountPojo=getUserAccountPojo();
         if(null==userAccountPojo ||  null==userAccountPojo.getId()){
             throw new BizException(ERRORCODE.USER_NO_EXIST.getCode(), ERRORCODE.USER_NO_EXIST.getMessage());
@@ -144,32 +161,9 @@ public class UserCollectController extends BaseController{
         Map<String,Object> param=Maps.newHashMap();
         param.put("userId",userId);
         param.put("projectId",projectId);
-        return userCollectExService.isUniversityCollect(param);
+        param.put("type",type);
+        return userCollectExService.isCollect(param);
     }
 
-    /**
-     * 查询指定用户收藏的所有课程
-     * @param offset
-     * @param rows
-     * @return
-     */
-    @RequestMapping(value = "/getUserCollectVideoCoursePojoList",method = RequestMethod.GET)
-    @ResponseBody
-    public Map<String,Object> getUserCollectVideoCoursePojoList(@RequestParam(value ="offset",required = false)Integer offset, @RequestParam(value = "rows",required = false)Integer rows){
-//        UserAccountPojo userAccountPojo=getUserAccountPojo();
-//        if(null==userAccountPojo ||  null==userAccountPojo.getId()){
-//            throw new BizException(ERRORCODE.USER_NO_EXIST.getCode(), ERRORCODE.USER_NO_EXIST.getMessage());
-//        }
-//        long userId=userAccountPojo.getId();
-        long userId=1L;
-        Map<String,Object> param= Maps.newHashMap();
-        param.put("userId",userId);
-        param.put("offset",offset);
-        param.put("rows",rows);
-        List<VideoCoursePojo> videoCoursePojoList=userCollectExService.getUserCollectVideoCoursePojoList(param);
-        Map<String,Object> returnParam=Maps.newHashMap();
-        returnParam.put("videoCoursePojoList",videoCoursePojoList);
-        return returnParam;
-    }
 
 }
